move the camera2d part to common
