+++ common/src/utils/camera.rs
@@ -1,8 +1,14 @@
-use glam::Mat4;
+use glam::{
+    Mat4,
+    Vec3,
+    Quat,
+    EulerRot,
+};
 
 use crate::{
     desktop::{InputEvent, WindowEvent},
     vr::{VrFov, VrUpdate},
+    Transform,
 };
 
 /// Perspective camera matrix utility
@@ -95,4 +101,95 @@ pub fn vr_projection_from_fov(fov: VrFov, near: f32, far: f32) -> Mat4 {
         [a31, a32, a33, -1.0],
         [0.0, 0.0, a43, 0.0],
     ])
+}
+
+pub struct Orthographic {
+    screen_size: (u32, u32),
+    pub left: f32,
+    pub right: f32,
+    pub bottom: f32,
+    pub top: f32,
+    pub near: f32,
+    pub far: f32,
+    proj: [Mat4; 2],
+}
+
+
+impl Default for Orthographic{
+    fn default() -> Self {
+        Self {
+            screen_size: (1920, 1080),
+            left: -10.,
+            right: 10.,
+            bottom: -10.,
+            top: 10.,
+            near: -100.,
+            far: 100.,
+            proj: [Mat4::IDENTITY; 2],
+        }
+    }
+}
+
+impl Orthographic{
+    pub fn update_proj(&mut self, width: f32, height: f32, input: &InputEvent) {
+        
+        // Check if the screen size changes
+        if let InputEvent::Window(WindowEvent::Resized { width: screen_width, height: screen_height }) = input {
+            self.screen_size = (*screen_width, *screen_height);
+        }
+
+        // Calculate the ratio of the screen size: need to work on this part again
+        let mut x_ratio = width;
+        let mut y_ratio = height;
+
+        while x_ratio / 10. >= 1. {
+            x_ratio /= 10.;
+        }
+
+        while y_ratio / 10. >= 1. {
+            y_ratio /= 10.;
+        }
+
+        // Update the ideal projection matrix of the screen
+        self.left = self.screen_size.0 as f32 / 2. / -(width / 2.) * x_ratio;
+        self.right = self.screen_size.0 as f32 / 2. / (width / 2.) *  x_ratio;
+        self.bottom = self.screen_size.1 as f32 / 2. / -(height / 2.) * y_ratio;
+        self.top = self.screen_size.1 as f32 / 2. / (height / 2.) * y_ratio;
+
+        dbg!(self.left, self.right, self.bottom, self.top);
+
+        // Recreate the new projection matrix based on the updated screen size        
+        let new_proj = Mat4::orthographic_rh_gl(
+            self.left,
+            self.right,
+            self.bottom,
+            self.top,
+            self.near,
+            self.far,
+        );
+        self.proj = [new_proj; 2];
+    }
+
+    pub fn new() -> Self {
+        Self::default()
+    }
+
+    pub fn matrices(&self) -> [Mat4; 2] {
+        self.proj
+    }
+
+    pub fn camera_on_positive_z_axis(&self) -> Transform {
+    // This camera control is for 2D arcade games that are played on a flat screen: z-axis is to us
+        Transform {
+            pos: Vec3::new(0., 0., 5.),
+            orient: Default::default(),
+        }
+    }
+
+    pub fn camera_on_custom_axis(&self, pos_x: f32, pos_y: f32, pos_z: f32, degree_x: f32, degree_y: f32, degree_z: f32) -> Transform {
+        Transform {
+            pos: Vec3::new(pos_x, pos_y, pos_z),
+            orient: Quat::from_euler(EulerRot::XYZ, degree_x.to_radians(), degree_y.to_radians(), degree_z.to_radians()),
+        }
+    }
 }
